refactor: share stream-json helpers
